sstable: add two more summary fields
After the keys array, the Summary file includes the first and last keys in this file's
range. Add this to the format.

Note that there is still more information after that. But that seems to be related to
the writing method (it says mmap in my files), and not relevant for us.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>