sstable summary: provide method to query an index in the summary
Although the entries are in an array, and live on disk, the disk array
abstraction turns out to be a bad abstraction to read it in. This is because,
contrary to other types, the key sizes are not to be found on-disk. It is a lot
more convenient to treat it as a normal array to be constructed as a separate
step.

We will construct this array at load time, and provide a method that, given an
index, returns the corresponding key/position.  After a binary search - to be
implemented - we'll be able to fetch the real data.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>