Test summary querying
Note: since the Summary files tend to contain very few indexes, all of the previous ones
had only a single entry. We are using small files so we don't have to carry files that are
too big in-tree.

To properly test the Summary files, though, I am importing from my Origin test
setup a quite large sstable, but zeroing its Data/Index file. We can use it to
test all metadata-related queries if we so want, but should not be expecting
actual data from it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>