sstables: fix use-after-free of temporary object
The debug-mode sanitizer discovered a bug in sstable::data_read.
I had optimistically wrote this code:

     return data_stream_at(pos).read_exactly(len);

This is wrong - data_stream_at returns a temporary input_stream object,
which gets destructed immediately and doesn't live throughout the life
of read_exactly. Obviously, this object does need to live on (among other
things, it holds the buffer which read_exactly reads into).

The solution is an ugly variant of the same thing, but which allocates
memory to hold a copy of the input stream object. Because there is no
single reader (in theory we can have a hundred different reads ongoing
in parallel from the same sstable), we really have choice but to allocate
this read context somewhere. A better solution would not use an input
stream at all, but this is a different issue, already in a FIXME.

This patch fixes the sstable test failure that Jenkins reports.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>