sstables: use do_with idiom
[v2: rebase, move "&", and tone down recommendation of template lambda]

Before commit f49c065cd16a75849eb4cca144d8d71ca866fa30, we had the buggy
but simple-looking code

    return data_stream_at(pos).read_exactly(len);

This was a bug because we need to ensure that the temporary object returned
by data_stream_at(pos) continues to live until the read_exactly() future
concludes. We solved this bug by doing this unsightly/unreadable code:

    auto stream = std::make_unique<input_stream<char>>(data_stream_at(pos));
    auto fut = stream->read_exactly(len);
    return fut.then([stream = std::move(stream)]
         (temporary_buffer<char> buf) { return buf; });

Instead, we can use the new do_with() idiom, which was exactly designed
to make a temporary object live until a future concludes. So we can write
the much shorter, and easier to understand, code:

    return do_with(data_stream_at(pos), [len] (auto &stream) {
        return stream.read_exactly(len);
    });

Note the C++14 template lambda (the "auto" in the argument of the lamda):
This lambda gets whatever we feed it (in this case, the stream returned
by data_stream_at). The "&" after the "auto" is important: without it,
the compiler tries to pass the object by value, which is impossible
(because it is not copyable).

Of course it would have also been possible to specify the stream's type
explicitly instead of using template lambda:

    return do_with(data_stream_at(pos), [len] (input_stream<char> &stream) {
        return stream.read_exactly(len);
    });

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>