sstables: test sstables::data_consume_row
Test the sstables::data_consume_row() function by consuming a row from
an sstable with known contents, and comparing the expected row key,
deletion time, and for each cell - its name, value and timestamp.

The test tests this on both a compressed and non-compressed sstable
(the contents of the two sstables is the same, except the timestamp
which is different).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>