sstables: do not short-cut cql row marker
As Nadav and Tomek both pointed out, the \0\0\0 is just a special case of a row
marker for a schema with no clustering keys. There is no need for us to handle
that separately.

And in fact, short-cutting it means that we will have to take care to update
the row timestamp in two places - which we weren't doing.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Reviewed-by: Nadav Har'El <nyh@cloudius-systems.com>